@@ -13,8 +13,9 @@
   ],
   "dependencies": {
     "@badrap/bar-of-progress": "^0.1.1",
-    "@docsearch/react": "^3.3.3",
-    "@headlessui/react": "^1.7.2",
+    "@docsearch/react": "^3.6.2",
+    "@headlessui/react": "^2.0.3",
+    "@heroicons/react": "^2.1.1",
     "@mdx-js/loader": "^2.3.0",
     "@mdx-js/react": "^2.3.0",
     "@silvenon/remark-smartypants": "^1.0.0",
@@ -31,6 +32,7 @@
     "debounce": "^1.2.0",
     "dlv": "^1.1.3",
     "exec-sh": "^0.3.4",
+    "fathom-client": "^3.7.0",
     "feed": "^4.2.2",
     "file-loader": "^6.0.0",
     "focus-visible": "^5.1.0",
@@ -59,25 +61,16 @@
     "seedrandom": "^3.0.5",
     "simple-functional-loader": "^1.2.1",
     "stringify-object": "^3.3.0",
-    "tinytime": "^0.2.6",
+    "tailwindcss": "^3.4.13",
     "unified": "^10.1.2",
     "unist-util-filter": "^4.0.1",
     "unist-util-visit": "^4.1.2",
     "zustand": "^4.0.0-rc.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "2.x",
-    "@typescript-eslint/parser": "2.x",
-    "babel-eslint": "10.x",
-    "eslint": "6.x",
-    "eslint-config-react-app": "^5.2.1",
-    "eslint-plugin-flowtype": "4.x",
-    "eslint-plugin-import": "2.x",
-    "eslint-plugin-jsx-a11y": "6.x",
-    "eslint-plugin-react": "7.x",
-    "eslint-plugin-react-hooks": "2.x",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^13.5.6",
     "glyphhanger": "^5.0.0",
-    "prettier": "^2.5.0",
-    "tailwindcss": "^3.3.3"
+    "prettier": "^2.5.0"
   }
 }
