@@ -26,7 +26,7 @@ export function Frame({ className, color = 'sky', children }) {
 export function EditorPane({ filename, scroll = false, code, children }) {
   return (
     <div className="pt-2 bg-slate-800 shadow-lg group">
-      <TabBar primary={{ name: filename }} showTabMarkers={false} />
+      {filename && <TabBar primary={{ name: filename }} showTabMarkers={false} />}
       <div
         className={clsx(
           'children:my-0 children:!shadow-none children:bg-transparent',
@@ -49,7 +49,7 @@ export function Editor({ filename, scroll = false, style = 'plain', color, child
 
   if (style === 'framed') {
     return (
-      <Frame className="mt-5 mb-8 first:mt-0 last:mb-0" color={color}>
+      <Frame className="mt-8 mb-8 first:mt-0 last:mb-0" color={color}>
         <EditorPane {...passthrough} filename={filename} code={code} children={children} />
       </Frame>
     )
