@@ -3,7 +3,7 @@ import { VersionSwitcher } from '@/components/VersionSwitcher'
 import { SearchButton } from '@/components/Search'
 import Router from 'next/router'
 import { Logo } from '@/components/Logo'
-import { Dialog } from '@headlessui/react'
+import { Dialog, DialogPanel } from '@headlessui/react'
 import { useEffect, useState } from 'react'
 import clsx from 'clsx'
 import { ThemeSelect, ThemeToggle } from './ThemeToggle'
@@ -11,10 +11,10 @@ import { ThemeSelect, ThemeToggle } from './ThemeToggle'
 function Featured() {
   return (
     <a
-      href="/blog/2023-07-18-tailwind-connect-2023-recap"
+      href="https://tailwindcss.com/blog/2024-05-24-catalyst-application-layouts"
       className="ml-3 text-xs leading-5 font-medium text-sky-600 dark:text-sky-400 bg-sky-400/10 rounded-full py-1 px-3 hidden xl:flex items-center hover:bg-sky-400/20"
     >
-      <strong className="font-semibold">Connect 2023 Recap</strong>
+      <strong className="font-semibold">Introducing Catalyst</strong>
       <svg
         width="2"
         height="2"
@@ -24,7 +24,7 @@ function Featured() {
       >
         <circle cx="1" cy="1" r="1" />
       </svg>
-      <span className="ml-2">Everything that was announced at our special event</span>
+      <span className="ml-2">A modern application UI kit for React</span>
       <svg
         width="3"
         height="6"
@@ -76,14 +76,9 @@ export function NavPopover({ display = 'md:hidden', className, ...props }) {
           />
         </svg>
       </button>
-      <Dialog
-        as="div"
-        className={clsx('fixed z-50 inset-0', display)}
-        open={isOpen}
-        onClose={setIsOpen}
-      >
-        <Dialog.Overlay className="fixed inset-0 bg-black/20 backdrop-blur-sm dark:bg-slate-900/80" />
-        <div className="fixed top-4 right-4 w-full max-w-xs bg-white rounded-lg shadow-lg p-6 text-base font-semibold text-slate-900 dark:bg-slate-800 dark:text-slate-400 dark:highlight-white/5">
+      <Dialog open={isOpen} onClose={setIsOpen} className={clsx('fixed z-50 inset-0', display)}>
+        <div className="fixed inset-0 bg-black/20 backdrop-blur-sm dark:bg-slate-900/80" />
+        <DialogPanel className="fixed top-4 right-4 w-full max-w-xs bg-white rounded-lg shadow-lg p-6 text-base font-semibold text-slate-900 dark:bg-slate-800 dark:text-slate-400 dark:highlight-white/5">
           <button
             type="button"
             className="absolute top-5 right-5 w-8 h-8 flex items-center justify-center text-slate-500 hover:text-slate-600 dark:text-slate-400 dark:hover:text-slate-300"
@@ -114,7 +109,7 @@ export function NavPopover({ display = 'md:hidden', className, ...props }) {
           <div className="mt-6 pt-6 border-t border-slate-200 dark:border-slate-200/10">
             <ThemeSelect />
           </div>
-        </div>
+        </DialogPanel>
       </Dialog>
     </div>
   )
