@@ -231,7 +231,7 @@ export function ContentsLayout({ children, meta, classes, tableOfContents, secti
         description={meta.description}
         repo={meta.repo}
         badge={{ key: 'Tailwind CSS version', value: meta.featureVersion }}
-        section={section}
+        section={section ?? meta.section}
       />
       <ContentsContext.Provider value={{ registerHeading, unregisterHeading }}>
         {classes ? (
