@@ -7,6 +7,12 @@ import { formatDate } from '@/utils/formatDate'
 import buildRss from '@/scripts/build-rss'
 import { renderToStaticMarkup } from 'react-dom/server'
 
+const dateFormat = {
+  month: 'long',
+  day: 'numeric',
+  year: 'numeric',
+}
+
 export default function Blog({ posts }) {
   return (
     <main className="max-w-[52rem] mx-auto px-4 pb-28 sm:px-6 md:px-8 xl:px-12 lg:max-w-6xl">
@@ -52,7 +58,7 @@ export default function Blog({ posts }) {
                 <dl className="absolute left-0 top-0 lg:left-auto lg:right-full lg:mr-[calc(6.5rem+1px)]">
                   <dt className="sr-only">Date</dt>
                   <dd className={clsx('whitespace-nowrap text-sm leading-6 dark:text-slate-400')}>
-                    <time dateTime={meta.date}>{formatDate(meta.date, '{MMMM} {DD}, {YYYY}')}</time>
+                    <time dateTime={meta.date}>{formatDate(meta.date, dateFormat)}</time>
                   </dd>
                 </dl>
               </div>
