@@ -115,7 +115,7 @@ let steps = [
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
->   <link href="./index.css" rel="stylesheet">
+>   <link href="./index.css" type="text/css" rel="stylesheet">
   </head>
   <body>
 >   <h1 class="text-3xl font-bold underline">
