@@ -58,7 +58,7 @@ let steps = [
     code: {
       name: 'Terminal',
       lang: 'terminal',
-      code: 'npx tailwindcss -i ./src/input.css -o ./dist/output.css --watch',
+      code: 'npx tailwindcss -i ./src/input.css -o ./src/output.css --watch',
     },
   },
   {
@@ -77,7 +77,7 @@ let steps = [
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
->   <link href="/dist/output.css" rel="stylesheet">
+>   <link href="./output.css" rel="stylesheet">
   </head>
   <body>
 >   <h1 class="text-3xl font-bold underline">
